Refactor recovery pipeline into staged executors
